# openapi.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "SWARM API",
    "description" : "API for the SWARM project. FAU CEN4214 2019F",
    "version" : "v0.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost:6969/api/v0",
    "description" : "Azure VM"
  } ],
  "paths" : {
...
    "/project" : {
      "get" : {
        "tags" : [ "Project" ],
        "description" : "Read information about a bunch of projects",
        "parameters" : [ {
          "name" : "offset",
          "in" : "query",
          "description" : "number of records to offset",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "max number of records to return",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "all of the projects",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "allOf" : [ {
                    "$ref" : "#/components/schemas/API-Response"
                  }, {
                    "properties" : {
                      "data" : {
                        "type" : "array",
                        "items" : {
                          "$ref" : "#/components/schemas/Project"
                        }
                      }
                    }
                  } ]
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Project" ],
        "description" : "Create a new devce",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Project"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "successfully registered project",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "allOf" : [ {
                    "$ref" : "#/components/schemas/API-Response"
                  }, {
                    "properties" : {
                      "data" : {
                        "$ref" : "#/components/schemas/Project"
                      }
                    }
                  } ]
                }
              }
            }
          }
        }
      }
    },
    "/project/{id}" : {
      "get" : {
        "tags" : [ "Project" ],
        "description" : "Read one project by ID",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "one of the projects",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "allOf" : [ {
                    "$ref" : "#/components/schemas/API-Response"
                  }, {
                    "properties" : {
                      "data" : {
                        "type" : "array",
                        "items" : {
                          "$ref" : "#/components/schemas/Project"
                        }
                      }
                    }
                  } ]
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Project" ],
        "description" : "Update one project by ID",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Project"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "successfully updated project",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "allOf" : [ {
                    "$ref" : "#/components/schemas/API-Response"
                  }, {
                    "properties" : {
                      "data" : {
                        "type" : "array",
                        "items" : {
                          "$ref" : "#/components/schemas/Project"
                        }
                      }
                    }
                  } ]
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Project" ],
        "description" : "Delete one project by ID",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successfully deleted project",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "allOf" : [ {
                    "$ref" : "#/components/schemas/API-Response"
                  }, {
                    "properties" : {
                      "data" : {
                        "type" : "array",
                        "items" : {
                          "$ref" : "#/components/schemas/Project"
                        }
                      }
                    }
                  } ]
                }
              }
            }
          }
        }
      }
    },
    "/user" : {
      "get" : {
        "tags" : [ "User" ],
        "description" : "Read information about a bunch of users",
        "parameters" : [ {
          "name" : "offset",
          "in" : "query",
          "description" : "number of records to offset",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "max number of records to return",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "all of the users",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "allOf" : [ {
                    "$ref" : "#/components/schemas/API-Response"
                  }, {
                    "properties" : {
                      "data" : {
                        "type" : "array",
                        "items" : {
                          "$ref" : "#/components/schemas/User"
                        }
                      }
                    }
                  } ]
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "User" ],
        "description" : "Create a new devce",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/User"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "successfully registered user",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "allOf" : [ {
                    "$ref" : "#/components/schemas/API-Response"
                  }, {
                    "properties" : {
                      "data" : {
                        "$ref" : "#/components/schemas/User"
                      }
                    }
                  } ]
                }
              }
            }
          }
        }
      }
    },
    "/user/{id}" : {
      "get" : {
        "tags" : [ "User" ],
        "description" : "Read one user by ID",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "one of the users",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "allOf" : [ {
                    "$ref" : "#/components/schemas/API-Response"
                  }, {
                    "properties" : {
                      "data" : {
                        "type" : "array",
                        "items" : {
                          "$ref" : "#/components/schemas/User"
                        }
                      }
                    }
                  } ]
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "User" ],
        "description" : "Update one user by ID",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/User"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "successfully updated user",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "allOf" : [ {
                    "$ref" : "#/components/schemas/API-Response"
                  }, {
                    "properties" : {
                      "data" : {
                        "type" : "array",
                        "items" : {
                          "$ref" : "#/components/schemas/User"
                        }
                      }
                    }
                  } ]
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "User" ],
        "description" : "Delete one user by ID",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successfully deleted user",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "allOf" : [ {
                    "$ref" : "#/components/schemas/API-Response"
                  }, {
                    "properties" : {
                      "data" : {
                        "type" : "array",
                        "items" : {
                          "$ref" : "#/components/schemas/User"
                        }
                      }
                    }
                  } ]
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "API-Response" : {
        "type" : "object",
        "properties" : {
          "message" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "data" : {
            "type" : "array",
            "items" : {
              "type" : "object"
            }
          },
          "error" : {
            "type" : "array",
            "items" : {
              "type" : "object"
            }
          }
        }
      },
      "Document" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "date_created" : {
            "type" : "string",
            "format" : "date-time"
          },
          "date_modified" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "Project" : {
        "type" : "object",
        "allOf" : [ {
          "$ref" : "#/components/schemas/Document"
        }, {
          "properties" : {
            "description" : {
              "type" : "string"
            },
            "img" : {
              "type" : "string"
            }
          }
        } ]
      },
      "User" : {
        "type" : "object",
        "allOf" : [ {
          "$ref" : "#/components/schemas/Document"
        }, {
          "properties" : {
            "first_name" : {
              "type" : "string"
            },
            "last_name" : {
              "type" : "string"
            },
            "email" : {
              "type" : "string",
              "format" : "email"
            },
            "password" : {
              "type" : "string"
            },
            "authorization" : {
              "type" : "object",
              "additionalProperties" : {
                "type" : "array",
                "items" : {
                  "type" : "string"
                }
              }
            }
          }
        } ]
      }
    }
  }
}